requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
python-socketio==5.10.0
eventlet==0.33.3
Werkzeug==3.0.1

# Load testing
requests==2.31.0
